+++ Src/FluentAssertions/Specialized/JsonNodeAssertions.cs
@@ -13,10 +13,11 @@ using FluentAssertions.Primitives;
 
 namespace FluentAssertions.Specialized;
 
-public class JsonNodeAssertions : ReferenceTypeAssertions<JsonNode, JsonNodeAssertions>
+public class JsonNodeAssertions<T> : ReferenceTypeAssertions<T, JsonNodeAssertions<T>>
+    where T : JsonNode
 {
     /// <inheritdoc />
-    public JsonNodeAssertions(JsonNode subject, AssertionChain assertionChain)
+    public JsonNodeAssertions(T subject, AssertionChain assertionChain)
         : base(subject, assertionChain)
     {
     }
@@ -38,7 +39,7 @@ public class JsonNodeAssertions : ReferenceTypeAssertions<JsonNode, JsonNodeAsse
     /// <param name="becauseArgs">
     /// Zero or more objects to format using the placeholders in <paramref name="because" />.
     /// </param>
-    public AndWhichConstraint<JsonNodeAssertions, JsonNode> HaveProperty(string code,
+    public AndWhichConstraint<JsonNodeAssertions<T>, JsonNode> HaveProperty(string code,
         [StringSyntax("CompositeFormat")] string because = "", params object[] becauseArgs)
     {
         CurrentAssertionChain
@@ -49,7 +50,7 @@ public class JsonNodeAssertions : ReferenceTypeAssertions<JsonNode, JsonNodeAsse
             .ForCondition(Subject?[code] != null)
             .FailWith("Expected {context:JSON node} to have property {0}{reason}.", code);
 
-        return new AndWhichConstraint<JsonNodeAssertions, JsonNode>(this, Subject?[code]);
+        return new AndWhichConstraint<JsonNodeAssertions<T>, JsonNode>(this, Subject?[code]);
     }
 
     /// <summary>
@@ -63,7 +64,7 @@ public class JsonNodeAssertions : ReferenceTypeAssertions<JsonNode, JsonNodeAsse
     /// <param name="becauseArgs">
     /// Zero or more objects to format using the placeholders in <paramref name="because" />.
     /// </param>
-    public AndConstraint<JsonNodeAssertions> NotHaveProperty(string code,
+    public AndConstraint<JsonNodeAssertions<T>> NotHaveProperty(string code,
         [StringSyntax("CompositeFormat")] string because = "", params object[] becauseArgs)
     {
         CurrentAssertionChain
@@ -75,7 +76,7 @@ public class JsonNodeAssertions : ReferenceTypeAssertions<JsonNode, JsonNodeAsse
             .BecauseOf(because, becauseArgs)
             .FailWith("Did not expect {context:JSON node} to have property {0}{reason}.", code);
 
-        return new AndConstraint<JsonNodeAssertions>(this);
+        return new AndConstraint<JsonNodeAssertions<T>>(this);
     }
 
     /// <summary>
@@ -88,7 +89,7 @@ public class JsonNodeAssertions : ReferenceTypeAssertions<JsonNode, JsonNodeAsse
     /// <param name="becauseArgs">
     /// Zero or more objects to format using the placeholders in <paramref name="because" />.
     /// </param>
-    public AndWhichConstraint<JsonNodeAssertions, IEnumerable<JsonNode>> BeAnArray(
+    public AndWhichConstraint<JsonNodeAssertions<T>, IEnumerable<JsonNode>> BeAnArray(
         [StringSyntax("CompositeFormat")] string because = "", params object[] becauseArgs)
     {
         CurrentAssertionChain
@@ -96,7 +97,7 @@ public class JsonNodeAssertions : ReferenceTypeAssertions<JsonNode, JsonNodeAsse
             .BecauseOf(because, becauseArgs)
             .FailWith("Expected {context:JSON node} to be an array{reason}, but {0} is not.", Subject);
 
-        return new AndWhichConstraint<JsonNodeAssertions, IEnumerable<JsonNode>>(this, (Subject as JsonArray)?.AsEnumerable());
+        return new AndWhichConstraint<JsonNodeAssertions<T>, IEnumerable<JsonNode>>(this, (Subject as JsonArray)?.AsEnumerable());
     }
 
     /// <summary>
@@ -109,7 +110,7 @@ public class JsonNodeAssertions : ReferenceTypeAssertions<JsonNode, JsonNodeAsse
     /// <param name="becauseArgs">
     /// Zero or more objects to format using the placeholders in <paramref name="because" />.
     /// </param>
-    public AndConstraint<JsonNodeAssertions> NotBeAnArray(
+    public AndConstraint<JsonNodeAssertions<T>> NotBeAnArray(
         [StringSyntax("CompositeFormat")] string because = "", params object[] becauseArgs)
     {
         CurrentAssertionChain
@@ -117,11 +118,11 @@ public class JsonNodeAssertions : ReferenceTypeAssertions<JsonNode, JsonNodeAsse
             .BecauseOf(because, becauseArgs)
             .FailWith("Did not expect {context:JSON node} to be an array{reason}, but {0} is.", Subject);
 
-        return new AndConstraint<JsonNodeAssertions>(this);
+        return new AndConstraint<JsonNodeAssertions<T>>(this);
     }
 
     /// <summary>
-    /// Asserts that the current <see cref="JsonNode"/> represents a 32-bit signed integer.
+    /// Asserts that the current <see cref="JsonNode"/> represents a numeric value.
     /// </summary>
     /// <param name="because">
     /// A formatted phrase as is supported by <see cref="string.Format(string,object[])" /> explaining why the assertion
@@ -130,20 +131,20 @@ public class JsonNodeAssertions : ReferenceTypeAssertions<JsonNode, JsonNodeAsse
     /// <param name="becauseArgs">
     /// Zero or more objects to format using the placeholders in <paramref name="because" />.
     /// </param>
-    public AndWhichConstraint<JsonNodeAssertions, long> BeNumeric(
+    public AndWhichConstraint<JsonNodeAssertions<T>, string> BeNumeric(
         [StringSyntax("CompositeFormat")] string because = "", params object[] becauseArgs)
     {
         CurrentAssertionChain
-            .ForCondition(Subject is JsonValue value && value.TryGetValue<int>(out _))
+            .ForCondition(Subject is JsonValue value && value.IsNumeric())
             .BecauseOf(because, becauseArgs)
-            .FailWith("Expected {context:JSON node} to be an Int32{reason}, but {0} is not.", Subject);
+            .FailWith("Expected {context:JSON node} to be a numeric value{reason}, but {0} is not.", Subject);
 
-        var actualValue = Subject is JsonValue jsonValue && jsonValue.TryGetValue<long>(out var result) ? result : 0;
-        return new AndWhichConstraint<JsonNodeAssertions, long>(this, actualValue);
+        return new AndWhichConstraint<JsonNodeAssertions<T>, string>(this, (Subject as JsonValue)?.ToString() ?? string.Empty);
     }
 
     /// <summary>
-    /// Asserts that the current <see cref="JsonNode"/> does not represent a 32-bit signed integer.
+    /// Asserts that the current <see cref="JsonNode"/> represents a numeric value and returns
+    /// the actual value provided it can be converted to <typeparamref name="TValue"/>.
     /// </summary>
     /// <param name="because">
     /// A formatted phrase as is supported by <see cref="string.Format(string,object[])" /> explaining why the assertion
@@ -152,15 +153,38 @@ public class JsonNodeAssertions : ReferenceTypeAssertions<JsonNode, JsonNodeAsse
     /// <param name="becauseArgs">
     /// Zero or more objects to format using the placeholders in <paramref name="because" />.
     /// </param>
-    public AndConstraint<JsonNodeAssertions> NotBeNumeric(
+    public AndWhichConstraint<JsonNodeAssertions<T>, TValue> BeNumeric<TValue>(
+        [StringSyntax("CompositeFormat")] string because = "", params object[] becauseArgs)
+    {
+        TValue actualValue = default;
+
+        CurrentAssertionChain
+            .ForCondition(Subject is JsonValue value && value.IsNumeric() && value.TryGetValue(out actualValue))
+            .BecauseOf(because, becauseArgs)
+            .FailWith("Expected {context:JSON node} to be a numeric value{reason}, but {0} is not.", Subject);
+
+        return new AndWhichConstraint<JsonNodeAssertions<T>, TValue>(this, actualValue);
+    }
+
+    /// <summary>
+    /// Asserts that the current <see cref="JsonNode"/> does not represent a numeric value.
+    /// </summary>
+    /// <param name="because">
+    /// A formatted phrase as is supported by <see cref="string.Format(string,object[])" /> explaining why the assertion
+    /// is needed. If the phrase does not start with the word <i>because</i>, it is prepended automatically.
+    /// </param>
+    /// <param name="becauseArgs">
+    /// Zero or more objects to format using the placeholders in <paramref name="because" />.
+    /// </param>
+    public AndConstraint<JsonNodeAssertions<T>> NotBeNumeric(
         [StringSyntax("CompositeFormat")] string because = "", params object[] becauseArgs)
     {
         CurrentAssertionChain
-            .ForCondition(!(Subject is JsonValue value && value.TryGetValue<long>(out _)))
+            .ForCondition(Subject is not JsonValue value || !value.IsNumeric())
             .BecauseOf(because, becauseArgs)
-            .FailWith("Did not expect {context:JSON node} to be an Int32{reason}, but {0} is.", Subject);
+            .FailWith("Did not expect {context:JSON node} to be a numeric value{reason}, but {0} is.", Subject);
 
-        return new AndConstraint<JsonNodeAssertions>(this);
+        return new AndConstraint<JsonNodeAssertions<T>>(this);
     }
 
     /// <summary>
@@ -173,7 +197,7 @@ public class JsonNodeAssertions : ReferenceTypeAssertions<JsonNode, JsonNodeAsse
     /// <param name="becauseArgs">
     /// Zero or more objects to format using the placeholders in <paramref name="because" />.
     /// </param>
-    public AndWhichConstraint<JsonNodeAssertions, DateTime> BeLocalDate(
+    public AndWhichConstraint<JsonNodeAssertions<T>, DateTime> BeLocalDate(
         [StringSyntax("CompositeFormat")] string because = "", params object[] becauseArgs)
     {
         CurrentAssertionChain
@@ -182,7 +206,7 @@ public class JsonNodeAssertions : ReferenceTypeAssertions<JsonNode, JsonNodeAsse
             .FailWith("Expected {context:JSON node} to be a local date{reason}, but {0} is not.", Subject);
 
         var actualValue = Subject is JsonValue jsonValue && jsonValue.TryGetValue<DateTime>(out var result) ? result : default;
-        return new AndWhichConstraint<JsonNodeAssertions, DateTime>(this, actualValue);
+        return new AndWhichConstraint<JsonNodeAssertions<T>, DateTime>(this, actualValue);
     }
 
     /// <summary>
@@ -195,15 +219,16 @@ public class JsonNodeAssertions : ReferenceTypeAssertions<JsonNode, JsonNodeAsse
     /// <param name="becauseArgs">
     /// Zero or more objects to format using the placeholders in <paramref name="because" />.
     /// </param>
-    public AndConstraint<JsonNodeAssertions> NotBeLocalDate(
+    public AndConstraint<JsonNodeAssertions<T>> NotBeLocalDate(
         [StringSyntax("CompositeFormat")] string because = "", params object[] becauseArgs)
     {
         CurrentAssertionChain
-            .ForCondition(Subject is not JsonValue value || !value.TryGetValue<DateTime>(out _) || Subject.ToString().EndsWith('Z'))
+            .ForCondition(Subject is not JsonValue value || !value.TryGetValue<DateTime>(out _) ||
+                          Subject.ToString().EndsWith('Z'))
             .BecauseOf(because, becauseArgs)
             .FailWith("Did not expect {context:JSON node} to be a local date{reason}, but {0} is.", Subject);
 
-        return new AndConstraint<JsonNodeAssertions>(this);
+        return new AndConstraint<JsonNodeAssertions<T>>(this);
     }
 
     /// <summary>
@@ -216,7 +241,7 @@ public class JsonNodeAssertions : ReferenceTypeAssertions<JsonNode, JsonNodeAsse
     /// <param name="becauseArgs">
     /// Zero or more objects to format using the placeholders in <paramref name="because" />.
     /// </param>
-    public AndWhichConstraint<JsonNodeAssertions, DateTime> BeUtcDate(
+    public AndWhichConstraint<JsonNodeAssertions<T>, DateTime> BeUtcDate(
         [StringSyntax("CompositeFormat")] string because = "", params object[] becauseArgs)
     {
         CurrentAssertionChain
@@ -225,7 +250,7 @@ public class JsonNodeAssertions : ReferenceTypeAssertions<JsonNode, JsonNodeAsse
             .FailWith("Expected {context} to be a UTC date{reason}, but {0} is not.", Subject);
 
         var actualValue = Subject is JsonValue jsonValue && jsonValue.TryGetValue<DateTime>(out var result) ? result : default;
-        return new AndWhichConstraint<JsonNodeAssertions, DateTime>(this, actualValue);
+        return new AndWhichConstraint<JsonNodeAssertions<T>, DateTime>(this, actualValue);
     }
 
     /// <summary>
@@ -238,15 +263,16 @@ public class JsonNodeAssertions : ReferenceTypeAssertions<JsonNode, JsonNodeAsse
     /// <param name="becauseArgs">
     /// Zero or more objects to format using the placeholders in <paramref name="because" />.
     /// </param>
-    public AndConstraint<JsonNodeAssertions> NotBeUtcDate(
+    public AndConstraint<JsonNodeAssertions<T>> NotBeUtcDate(
         [StringSyntax("CompositeFormat")] string because = "", params object[] becauseArgs)
     {
         CurrentAssertionChain
-            .ForCondition(Subject is not JsonValue value || !value.TryGetValue<DateTime>(out _) || !Subject.ToString().EndsWith('Z'))
+            .ForCondition(Subject is not JsonValue value || !value.TryGetValue<DateTime>(out _) ||
+                          !Subject.ToString().EndsWith('Z'))
             .BecauseOf(because, becauseArgs)
             .FailWith("Did not expect {context} to be a UTC date{reason}, but {0} is.", Subject);
 
-        return new AndConstraint<JsonNodeAssertions>(this);
+        return new AndConstraint<JsonNodeAssertions<T>>(this);
     }
 
     /// <summary>
@@ -259,7 +285,7 @@ public class JsonNodeAssertions : ReferenceTypeAssertions<JsonNode, JsonNodeAsse
     /// <param name="becauseArgs">
     /// Zero or more objects to format using the placeholders in <paramref name="because" />.
     /// </param>
-    public AndWhichConstraint<JsonNodeAssertions, bool> BeBool(
+    public AndWhichConstraint<JsonNodeAssertions<T>, bool> BeBool(
         [StringSyntax("CompositeFormat")] string because = "", params object[] becauseArgs)
     {
         CurrentAssertionChain
@@ -268,7 +294,7 @@ public class JsonNodeAssertions : ReferenceTypeAssertions<JsonNode, JsonNodeAsse
             .FailWith("Expected {context} to be a boolean{reason}, but {0} is not.", Subject);
 
         var actualValue = Subject is JsonValue jsonValue && jsonValue.TryGetValue<bool>(out var result) && result;
-        return new AndWhichConstraint<JsonNodeAssertions, bool>(this, actualValue);
+        return new AndWhichConstraint<JsonNodeAssertions<T>, bool>(this, actualValue);
     }
 
     /// <summary>
@@ -281,7 +307,7 @@ public class JsonNodeAssertions : ReferenceTypeAssertions<JsonNode, JsonNodeAsse
     /// <param name="becauseArgs">
     /// Zero or more objects to format using the placeholders in <paramref name="because" />.
     /// </param>
-    public AndConstraint<JsonNodeAssertions> NotBeBool(
+    public AndConstraint<JsonNodeAssertions<T>> NotBeBool(
         [StringSyntax("CompositeFormat")] string because = "", params object[] becauseArgs)
     {
         CurrentAssertionChain
@@ -289,7 +315,7 @@ public class JsonNodeAssertions : ReferenceTypeAssertions<JsonNode, JsonNodeAsse
             .BecauseOf(because, becauseArgs)
             .FailWith("Did not expect {context:JSON node} to be a boolean{reason}, but {0} is.", Subject);
 
-        return new AndConstraint<JsonNodeAssertions>(this);
+        return new AndConstraint<JsonNodeAssertions<T>>(this);
     }
 
     /// <summary>
@@ -302,7 +328,7 @@ public class JsonNodeAssertions : ReferenceTypeAssertions<JsonNode, JsonNodeAsse
     /// <param name="becauseArgs">
     /// Zero or more objects to format using the placeholders in <paramref name="because" />.
     /// </param>
-    public AndWhichConstraint<JsonNodeAssertions, string> BeString(
+    public AndWhichConstraint<JsonNodeAssertions<T>, string> BeString(
         [StringSyntax("CompositeFormat")] string because = "", params object[] becauseArgs)
     {
         CurrentAssertionChain
@@ -311,7 +337,7 @@ public class JsonNodeAssertions : ReferenceTypeAssertions<JsonNode, JsonNodeAsse
             .FailWith("Expected {context:JSON node} to be a string{reason}, but {0} is not.", Subject);
 
         var actualValue = Subject is JsonValue jsonValue && jsonValue.TryGetValue<string>(out var result) ? result : null;
-        return new AndWhichConstraint<JsonNodeAssertions, string>(this, actualValue);
+        return new AndWhichConstraint<JsonNodeAssertions<T>, string>(this, actualValue);
     }
 
     /// <summary>
@@ -324,7 +350,7 @@ public class JsonNodeAssertions : ReferenceTypeAssertions<JsonNode, JsonNodeAsse
     /// <param name="becauseArgs">
     /// Zero or more objects to format using the placeholders in <paramref name="because" />.
     /// </param>
-    public AndConstraint<JsonNodeAssertions> NotBeString(
+    public AndConstraint<JsonNodeAssertions<T>> NotBeString(
         [StringSyntax("CompositeFormat")] string because = "", params object[] becauseArgs)
     {
         CurrentAssertionChain
@@ -332,7 +358,7 @@ public class JsonNodeAssertions : ReferenceTypeAssertions<JsonNode, JsonNodeAsse
             .BecauseOf(because, becauseArgs)
             .FailWith("Did not expect {context:JSON node} to be a string{reason}, but {0} is.", Subject);
 
-        return new AndConstraint<JsonNodeAssertions>(this);
+        return new AndConstraint<JsonNodeAssertions<T>>(this);
     }
 
     /// <summary>
@@ -355,7 +381,7 @@ public class JsonNodeAssertions : ReferenceTypeAssertions<JsonNode, JsonNodeAsse
     /// <param name="becauseArgs">
     /// Zero or more objects to format using the placeholders in <paramref name="because"/>.
     /// </param>
-    public AndConstraint<JsonNodeAssertions> BeEquivalentTo<TExpectation>(TExpectation expectation,
+    public AndConstraint<JsonNodeAssertions<T>> BeEquivalentTo<TExpectation>(TExpectation expectation,
         [StringSyntax("CompositeFormat")] string because = "", params object[] becauseArgs)
     {
         return BeEquivalentTo(expectation, config => config, because, becauseArgs);
@@ -389,7 +415,7 @@ public class JsonNodeAssertions : ReferenceTypeAssertions<JsonNode, JsonNodeAsse
     /// <param name="becauseArgs">
     /// Zero or more objects to format using the placeholders in <paramref name="because"/>.
     /// </param>
-    public AndConstraint<JsonNodeAssertions> BeEquivalentTo<TExpectation>(TExpectation expectation,
+    public AndConstraint<JsonNodeAssertions<T>> BeEquivalentTo<TExpectation>(TExpectation expectation,
         Func<EquivalencyOptions<TExpectation>, EquivalencyOptions<TExpectation>> config,
         [StringSyntax("CompositeFormat")] string because = "", params object[] becauseArgs)
     {
@@ -414,7 +440,7 @@ public class JsonNodeAssertions : ReferenceTypeAssertions<JsonNode, JsonNodeAsse
 
         new EquivalencyValidator().AssertEquality(comparands, context);
 
-        return new AndConstraint<JsonNodeAssertions>(this);
+        return new AndConstraint<JsonNodeAssertions<T>>(this);
     }
 
     /// <summary>
@@ -437,7 +463,7 @@ public class JsonNodeAssertions : ReferenceTypeAssertions<JsonNode, JsonNodeAsse
     /// <param name="becauseArgs">
     /// Zero or more objects to format using the placeholders in <paramref name="because"/>.
     /// </param>
-    public AndConstraint<JsonNodeAssertions> NotBeEquivalentTo<TExpectation>(
+    public AndConstraint<JsonNodeAssertions<T>> NotBeEquivalentTo<TExpectation>(
         TExpectation unexpected,
         [StringSyntax("CompositeFormat")] string because = "", params object[] becauseArgs)
     {
@@ -472,7 +498,7 @@ public class JsonNodeAssertions : ReferenceTypeAssertions<JsonNode, JsonNodeAsse
     /// <param name="becauseArgs">
     /// Zero or more objects to format using the placeholders in <paramref name="because"/>.
     /// </param>
-    public AndConstraint<JsonNodeAssertions> NotBeEquivalentTo<TExpectation>(
+    public AndConstraint<JsonNodeAssertions<T>> NotBeEquivalentTo<TExpectation>(
         TExpectation unexpected,
         Func<EquivalencyOptions<TExpectation>, EquivalencyOptions<TExpectation>> config,
         [StringSyntax("CompositeFormat")] string because = "", params object[] becauseArgs)
@@ -493,7 +519,7 @@ public class JsonNodeAssertions : ReferenceTypeAssertions<JsonNode, JsonNodeAsse
             .WithDefaultIdentifier(Identifier)
             .FailWith("Did not expect {context:JSON node} to be equivalent to {0}{reason}, but they are.", unexpected);
 
-        return new AndConstraint<JsonNodeAssertions>(this);
+        return new AndConstraint<JsonNodeAssertions<T>>(this);
     }
 }