+++ docs/_pages/releases.md
@@ -7,6 +7,13 @@ sidebar:
   nav: "sidebar"
 ---
 
+## 8.7.1
+
+### Fixes
+
+* Fixed ambiguity when using `Should()` on a `JsonNode` derived class - [#3102](https://github.com/fluentassertions/fluentassertions/pull/3102)
+* JSON assertions did not properly handle floats, doubles and unsigned int/long - [#3105](https://github.com/fluentassertions/fluentassertions/pull/3105)
+
 ## 8.7.0
 
 ### What's new